adicionado mais 12 seeders com titulo e fotos, de acordo com a pagina da loja

+++ database/migrations/20220512120649-produtos.js
@@ -19,7 +19,7 @@ module.exports = {
         allowNull: false
       },            
       preco: {
-        type: Sequelize.DataTypes.INTEGER,
+        type: Sequelize.DataTypes.DECIMAL(10,2),
         allowNull: false
       },
       rating: {
